Restructure guess code
